% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefs.R
\name{get_coefs}
\alias{get_coefs}
\title{Get coeficient names from a list of fitted models.}
\usage{
get_coefs(models)
}
\arguments{
\item{models}{list of model objects made by \code{get_fit_models}.}

\item{data}{list having the data used for training and testing the 
various model types.}
}
\value{
list character vectors with coeficient names named with the name
of the variable selection (reduction) technique used.
}
\description{
Get coeficient names from a list of fitted models.
}
