% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cv_glmnet.R
\name{get_cv_glmnet}
\alias{get_cv_glmnet}
\title{Wrapper around cv.glmnet}
\usage{
get_cv_glmnet(x, y, cv_alpha)
}
\arguments{
\item{x}{matrix of predictor varibles}

\item{y}{vector of target variables}

\item{cv_alpha}{The elasticnet mixing parameter}
}
\description{
Wrapper around cv.glmnet
}
