% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_families.R
\name{get_families}
\alias{get_families}
\title{Get a list of the distribution families used from a list of fitted model objects}
\usage{
get_families(models)
}
\arguments{
\item{models}{list of fitted models}
}
\value{
character vector of the distribution families used to fit the model
}
\description{
Get a list of the distribution families used from a list of fitted model objects
}
