% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_models.R
\name{get_fit_models}
\alias{get_fit_models}
\title{Fits list of models to provided data}
\usage{
get_fit_models(data, models, ...)
}
\arguments{
\item{data}{data set made by \code{make_dataset}}

\item{models}{character vector of model types to fit. Can be
\code{"backward"}, \code{"forward"}, \code{"both"}, \code{"ridge"}
\code{"lasso"}, or \code{"elastic_net"}.}

\item{...}{additional named parameters needed to fit the models.}
}
\value{
List of fitted model objects
}
\description{
Fits list of models to provided data
}
