% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fivenum_coef.R
\name{get_fivenum_coef}
\alias{get_fivenum_coef}
\title{Gets fivenum() summary for coefficients retained by various models}
\usage{
get_fivenum_coef(results, authentic, noise)
}
\arguments{
\item{results}{list of results created inside simulation loop}

\item{authentic}{character vector of authentic predictive variables.}

\item{noise}{character vector of noise predictive variables.}
}
\description{
Gets fivenum() summary for coefficients retained by various models
}
