% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fivenum_mse.R
\name{get_fivenum_mse}
\alias{get_fivenum_mse}
\title{Get fivenum summary of mean squared error from results list}
\usage{
get_fivenum_mse(results)
}
\arguments{
\item{results}{list of results from regressions}
}
\description{
Get fivenum summary of mean squared error from results list
}
