% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glmnet_coef.R
\name{get_glmnet_coef}
\alias{get_glmnet_coef}
\title{Get coeficients without intercept from glmnet model}
\usage{
get_glmnet_coef(fit)
}
\arguments{
\item{fit}{glmnet model}
}
\description{
Get coeficients without intercept from glmnet model
}
