% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links}
\alias{get_links}
\title{Get a list of the link functions used from a list of fitted model objects}
\usage{
get_links(models)
}
\arguments{
\item{models}{list of fitted models}
}
\value{
character vector of the link functions used to fit the models
}
\description{
This function does not really work for glmnet. There is not link function
information directly stored within the fitted model.
}
