% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mses.R
\name{get_mses}
\alias{get_mses}
\title{Get Mean Squared Errors for a list of models using test data.}
\usage{
get_mses(models, data)
}
\arguments{
\item{models}{list of model objects made by \code{get_fit_models}.}

\item{data}{list having the data used for training and testing the 
various model types.}
}
\value{
list object with Mean Squared Errors named with the name
of the variable selection (reduction) technique used.
}
\description{
Get Mean Squared Errors for a list of models using test data.
}
