% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictors.R
\name{get_predictors}
\alias{get_predictors}
\title{get_predictors returns matrix of predictor variable data elements}
\usage{
get_predictors(y, betas, corr, p, error_fun)
}
\arguments{
\item{y}{target value observations}

\item{betas}{coefficients to use to make the target values}

\item{corr}{correlation parameters that is a vector of length p. The X 
columns created with a correlation parameter value greater than 0
are authentic predictors and those with no correlation parameter are noise}

\item{p}{number of predictive variables}

\item{error_fun}{error function}
}
\value{
matrix of continuous predictive variables
}
\description{
get_predictors returns matrix of predictor variable data elements
}
