% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_targets.R
\name{get_targets}
\alias{get_targets}
\title{get_targets returns matrix of target data elements}
\usage{
get_targets(betas, X, error_fun, ...)
}
\arguments{
\item{betas}{coefficients to use to make the target values}

\item{X}{a matrix of observed predictor variable values}

\item{error_fun}{error function}

\item{...}{additional arguments need by \code{error_fun}}
}
\value{
matrix of target variables
}
\description{
get_targets returns matrix of target data elements
}
