% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataset.R
\name{make_dataset}
\alias{make_dataset}
\title{Forms data set used for simulation}
\usage{
make_dataset(n = 100, betas, corr, p, target_mean = 0,
  target_sd = 100, error_fun_mean = 1:10, error_fun_sd = 50:200)
}
\arguments{
\item{n}{number of observations}

\item{betas}{coefficients used to form actual predictive variables}

\item{corr}{correlations used in making predictive variables}

\item{p}{number of predictive variables}

\item{target_mean}{mean value of target variable}

\item{target_sd}{standard deviation of target variable}

\item{error_fun_mean}{vector of possible means used in the error function 
\code{error_fun}.}

\item{error_fun_sd}{vector of possible standard deviations used in the error function
\code{error_fun}}
}
\description{
The data set is
}
