% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results.R
\name{make_results}
\alias{make_results}
\title{Makes a result list object with all simulation results included.}
\usage{
make_results(direction, n, alpha, rho, p, models, data, sim)
}
\arguments{
\item{direction}{character vector with one of "backward", "forward",
or "both".}

\item{n}{sample size}

\item{alpha}{threshold value}

\item{rho}{correlations between real predictors}

\item{p}{number of predictor variables to include}

\item{models}{list of fitted models}

\item{data}{list having the data used for training and testing the 
various model types.}

\item{sim}{number of simulations}
}
\value{
The most complete form of the results of the simulation
are provided by this object. It has a sublist of results for every
simulated sample.
}
\description{
Makes a result list object with all simulation results included.
}
