% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_object_to_df.R
\name{sim_object_to_df}
\alias{sim_object_to_df}
\title{sim_object_to_df creates a dataframe from the simulation object created by 
\code{dk_sim}}
\usage{
sim_object_to_df(sim_object)
}
\arguments{
\item{sim_object}{list object made by \code{dk_sim} function}
}
\value{
dataframe from the simulation object created by \code{dk_sim}
function
}
\description{
sim_object_to_df creates a dataframe from the simulation object created by 
\code{dk_sim}
}
