% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictors.R
\name{get_predictors}
\alias{get_predictors}
\title{get_predictors returns matrix of predictor variable data elements}
\usage{
get_predictors(corr, n, p, Z)
}
\arguments{
\item{corr}{correlation parameters that is a vector of length p. The X 
columns created with a correlation parameter value greater than 0
are authentic predictors and those with}

\item{n}{number of observations (subjects)}

\item{p}{number of predictive variables}

\item{Z}{a matrix of independent identically distributed pseudorandom 
variates with \code{n} rows and \code{p + 1} columns.}
}
\value{
matrix of continuous predictive variables
}
\description{
get_predictors returns matrix of predictor variable data elements
}
