% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataset.R
\name{make_dataset}
\alias{make_dataset}
\title{Forms dataset used for simulation}
\usage{
make_dataset(n, betas, corr, p)
}
\arguments{
\item{n}{number of observations}

\item{betas}{coefficients used to form actual predictive variables}

\item{corr}{correlations used in making predictive variables}

\item{p}{number of predictive variables}
}
\description{
Forms dataset used for simulation
}
